Translate to spanish/add customStyles/add regular exp

+++ src/app/components/backoffices/admin/tenants-list/tenants-list.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { TenantsService } from '../../../../services/admin/tenants.service';
-
+import Swal from 'sweetalert2';
 @Component({
   selector: 'app-tenants-list',
   templateUrl: './tenants-list.component.html',
@@ -28,14 +28,31 @@ export class TenantsListComponent implements OnInit {
   }
 
   async deleteTenant(id: any) {
-    const {ok, message, data}:any = await this.tenantsService.deleteTenant(id);
-    console.log(message);
-    if(!ok) {
-      console.log(`${message}`);
-      console.log(`Tenants couldn't delete.`);
-    }
-    const { tenant } = data;
-    this.readTenants();
+    Swal.fire({
+      title: 'Estas seguro de eliminar el consultorio?',
+      text: "Todos los datos asociados al consultorio seran eliminados, no podras revertir esto",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Borrar'
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        const {ok, message, data}:any = await this.tenantsService.deleteTenant(id);
+        console.log(message);
+        if(!ok) {
+          console.log(`${message}`);
+          console.log(`Tenants couldn't delete.`);
+        }
+        const { tenant } = data;
+        this.readTenants();
+        Swal.fire(
+          'Eliminado!',
+          'Consultorio eliminado con exito.',
+          'success'
+        )
+      }
+    })
   }
 
 }
