+++ src/app/components/backoffices/operator/operator-appointments/operator-appointments.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 
 import { OperatorAppointmentsService } from '../../../../services/operator/operator-appointments.service';
 
+import Swal from 'sweetalert2';
+
 @Component({
   selector: 'app-operator-appointments',
   templateUrl: './operator-appointments.component.html',
@@ -26,8 +28,24 @@ export class OperatorAppointmentsComponent implements OnInit {
   }
 
   async deleteAppointment(apptId:any) {
-    const {appointment}:any = await this.operatorAppointmentsService.deleteAppointment(apptId);
-    this.listOperatorlAppointments();
+    Swal.fire({
+      title: 'Estas seguro de eliminar turno?',
+      text: "No podras revertir los cambios",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Borrar'
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        const {appointment}:any = await this.operatorAppointmentsService.deleteAppointment(apptId);
+        this.listOperatorlAppointments();
+        Swal.fire(
+          'Eliminado!',
+          'Turno eliminado con exito',
+          'success'
+        )
+      }
+    })
   }
-
 }
