+++ src/app/components/backoffices/patient/patient-appointments/patient-appointments.component.ts
@@ -11,9 +11,7 @@ import Swal from 'sweetalert2';
 export class PatientAppointmentsComponent implements OnInit {
 
   appointments: any[] = [];
-  // tenants: any[] = [];
-  // selectedTenant: any;
-  // selectedEspecialidad: any;
+
 
   constructor(
     private patientAppointmentsService:PatientAppointmentsService,
@@ -30,13 +28,28 @@ export class PatientAppointmentsComponent implements OnInit {
   }
 
   async deleteAppointment(apptId:any) {
-    const { ok, appointmentId } = await this.patientAppointmentsService.deleteAppointment(apptId);
-    if(ok) {
-      this.listPatientAppointments();
-    } else {
-      console.log(`Error to delete the appointment: ${apptId}`);
-    }
+    Swal.fire({
+      title: 'Estas seguro de eliminar el turno?',
+      text: "No podras revertir los cambios",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Borrar'
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        const { ok, appointmentId } = await this.patientAppointmentsService.deleteAppointment(apptId);
+        if(ok) {
+          this.listPatientAppointments();
+        } else {
+          console.log(`Error to delete the appointment: ${apptId}`);
+        }
+        Swal.fire(
+          'Eliminado',
+          'Turno eliminado con exito',
+          'success'
+        )
+      }
+    })
   }
-
-
 }
